Updating to 1.3
