+++ .gitignore
@@ -1,0 +1,8 @@
+
+*.lock
+*.ide
+*.ide-shm
+*.ide-wal
+1.1/Source/.vs/Regeneration/v15/.suo
+1.1/Source/.vs/Regeneration/v16/.suo
+1.3/Source/Regeneration/.vs/Regeneration/v16/.suo
