+++ 1.1/Defs/HediffDefs/RegenHediffs.xml
