+++ 1.1/Defs/ResearchProjectDefs/ResearchProjects_Regeneration.xml
