+++ 1.1/Defs/ThoughtDefs/Thoughts_Regeneration.xml
