+++ 1.1/Source/Regeneration/ModSettings.cs
