+++ 1.1/Source/Regeneration/Properties/AssemblyInfo.cs
