+++ 1.1/Source/Regeneration/Regeneration.csproj
