+++ 1.3/Defs/ThoughtDefs/Thoughts_Regeneration.xml
@@ -1,0 +1,66 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<Defs>
+  <ThoughtDef ParentName="DeathMemory">
+    <defName>RecentlyRegenerated</defName>
+    <durationDays>3</durationDays>
+    <stackLimit>5</stackLimit>
+    <nullifyingTraits>
+      <li>Psychopath</li>
+      <li>Bloodlust</li>
+    </nullifyingTraits>
+    <stages>
+      <li>
+        <label>Recently resurrected</label>
+        <description>I recently resurrected. I'm not sure who I am anymore.</description>
+        <baseMoodEffect>-3</baseMoodEffect>
+      </li>
+    </stages>
+  </ThoughtDef>
+  
+  <ThoughtDef ParentName="DeathMemory">
+    <defName>AncientRecentlyRegenerated</defName>
+    <durationDays>60</durationDays>
+    <stackLimit>100</stackLimit>
+    <nullifyingTraits>
+      <li>Psychopath</li>
+      <li>Bloodlust</li>
+    </nullifyingTraits>
+    <stages>
+      <li>
+        <label>where is death?</label>
+        <description>I resurrected... again. Where is death?</description>
+        <baseMoodEffect>-5</baseMoodEffect>
+      </li>
+    </stages>
+  </ThoughtDef>
+  
+  <ThoughtDef>
+    <defName>KnownColonistRegeneratedSocial</defName>
+    <thoughtClass>Thought_MemorySocial</thoughtClass>
+    <durationDays>3</durationDays>
+    <stackLimit>12</stackLimit>
+    <stackLimitForSameOtherPawn>12</stackLimitForSameOtherPawn>
+    <stackedEffectMultiplier>0.9</stackedEffectMultiplier>
+    <stages>
+      <li>
+        <label>family member regenerated</label>
+        <baseOpinionOffset>-15</baseOpinionOffset>
+      </li>
+    </stages>
+  </ThoughtDef>
+  
+  <ThoughtDef>
+    <defName>KnownColonistRegenerated</defName>
+    <durationDays>3</durationDays>
+    <stackLimit>12</stackLimit>
+    <stackedEffectMultiplier>0.9</stackedEffectMultiplier>
+    <stages>
+      <li>
+        <label>family member resurrected</label>
+        <description>One of my relations has changed through resurrection. I'm not sure if this new incarnation can be trusted.</description>
+        <baseMoodEffect>-5</baseMoodEffect>
+      </li>
+    </stages>
+  </ThoughtDef>
+  
+</Defs>
