+++ 1.3/Source/Regeneration/Regeneration.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31613.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Regeneration", "Regeneration.csproj", "{9AD6156A-A8DB-435F-8FDB-70439E84E3BD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9AD6156A-A8DB-435F-8FDB-70439E84E3BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9AD6156A-A8DB-435F-8FDB-70439E84E3BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9AD6156A-A8DB-435F-8FDB-70439E84E3BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9AD6156A-A8DB-435F-8FDB-70439E84E3BD}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4BF92234-C5FD-4CE0-8C34-DD9048C631D7}
+	EndGlobalSection
+EndGlobal
