+++ 1.3/Source/Regeneration/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Lib.Harmony" version="2.1.1" targetFramework="net472" />
+</packages>
