+++ About/About.xml
@@ -1,11 +1,12 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <ModMetaData>
-	<name>Rimmortality (WIP)</name>
-	<author>Demeggy</author>
-	 <supportedVersions>
-      <li>1.1</li>
-   </supportedVersions>
-   <packageId>demeggy.rimmortality</packageId>
-	<description>Auto-resurrect upon death, but randomise appearance and traits, and give debuff to all related colonists
-</description>
+	<name>Rimmortality</name>
+	<author>Demeggy, Primuth</author>
+		<supportedVersions>
+		<li>1.1</li>
+		<li>1.3</li>
+	</supportedVersions>
+	<packageId>demeggy.rimmortality</packageId>
+	<description>Auto-resurrect upon death, but randomise appearance and traits, and give debuff to all related colonists. Inspired by the regeneration process in Doctor Who.
+	</description>
 </ModMetaData>
