+++ README.md
@@ -1,0 +1,3 @@
+## Rimmortality Mod
+
+Inspired by Doctor Who, this mod allows you to inject your pawns with a serum granting the ability for twelve regenerations into random pawns, triggered upon death.
